Решение максимальной площади воды
